feat(inbox): tab title and favicon update on new message
