+++ index.html
@@ -7,14 +7,14 @@
     <meta name="description" content="" />
     <meta name="robots" content="noindex" />
     <title>Ayon</title>
-    <!-- FAVICON -->
+    <!-- FALLBACK FAVICON -->
     <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png">
     <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png">
     <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png">
     <link rel="manifest" href="/site.webmanifest">
     <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#00d7a0">
     <meta name="msapplication-TileColor" content="#00aba9">
-    <!-- FAVICON -->
+    <!-- FALLBACK FAVICON -->
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
