+++ package.json
@@ -32,6 +32,7 @@
     "react-datepicker": "^4.11.0",
     "react-dom": "^18.0.0",
     "react-error-boundary": "^4.0.3",
+    "react-helmet": "^6.1.0",
     "react-intersection-observer": "^9.5.2",
     "react-markdown": "^9.0.1",
     "react-perfect-scrollbar": "^1.5.8",
