+++ public/site-new.webmanifest
@@ -1,0 +1,19 @@
+{
+  "name": "AYON",
+  "short_name": "AYON",
+  "icons": [
+    {
+      "src": "/android-chrome-192x192-notification.png",
+      "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "/android-chrome-512x512-notification.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    }
+  ],
+  "theme_color": "#2c313a",
+  "background_color": "#2c313a",
+  "display": "standalone"
+}
