+++ src/containers/header/AppHeader.jsx
@@ -18,7 +18,7 @@ import { toggleDevMode } from '/src/features/user'
 import styled from 'styled-components'
 import { useRestart } from '/src/context/restartContext'
 import { classNames } from 'primereact/utils'
-import { useGetInboxHasUnreadQuery } from '/src/services/inbox/getInbox'
+import InboxNotificationIcon from './InboxNotificationIcon'
 
 const DeveloperSwitch = styled.div`
   display: flex;
@@ -107,23 +107,6 @@ const Header = () => {
     }
   }, [location.pathname, localStorage])
 
-  const { data: isNewMessages, refetch } = useGetInboxHasUnreadQuery()
-
-  useEffect(() => {
-    refetch() // Check messages immediately on location change
-
-    const interval = setInterval(refetch, 600000) // Check messages every 10 minutes
-
-    const timeout = setTimeout(() => {
-      refetch() // Check messages after 10 minutes, even if the location hasn't changed
-    }, 600000)
-
-    return () => {
-      clearInterval(interval)
-      clearTimeout(timeout)
-    }
-  }, [location.pathname])
-
   const handleNavClick = (e) => {
     // if target us nav, then close menu
     if (e.target.tagName === 'NAV') handleSetMenu(false)
@@ -202,13 +185,7 @@ const Header = () => {
       {/* help icon and menu ^^^ */}
 
       {/* Inbox icon */}
-      <Link to="/inbox/important">
-        <HeaderButton
-          icon="inbox"
-          variant="nav"
-          className={classNames({ notification: isNewMessages })}
-        />
-      </Link>
+      <InboxNotificationIcon />
 
       {/* App icon and menu vvv */}
       <HeaderButton
